+++ rs-week1/src/__tests__/Header.test.tsx
@@ -1,27 +1,29 @@
-import { describe, test, expect } from 'vitest';
 import React from 'react';
+import { render, screen } from '@testing-library/react';
 import Header from '../components/Header';
-import { render, fireEvent } from '@testing-library/react';
-import { fn } from 'jest-mock';
 import { MemoryRouter } from 'react-router';
 
-describe('<Header />', () => {
-  test('handles search input change and form submission', () => {
-    const onSearch = fn();
-    const { getByPlaceholderText, getByText } = render(
-      <MemoryRouter>
-        <Header onSearch={onSearch} namePageTerm="" />
-      </MemoryRouter>
-    );
-    const searchInput = getByPlaceholderText('Search...') as HTMLInputElement;
-    expect(searchInput).toBeTruthy();
-    const searchButton = getByText('Search');
+test('should render the component with the correct namePageTerm prop', () => {
+  const namePageTerm = 'Home Page';
+  render(
+    <MemoryRouter>
+      <Header namePageTerm={namePageTerm} />
+    </MemoryRouter>
+  );
 
-    fireEvent.change(searchInput, { target: { value: 'test' } });
-    expect(searchInput.value).toBe('test');
+  expect(screen.getByText(`Current Page: ${namePageTerm}`)).toBeTruthy();
+});
 
-    fireEvent.click(searchButton);
-    expect(onSearch).toHaveBeenCalledTimes(1);
-    expect(onSearch).toHaveBeenCalledWith('test');
-  });
+test('should render the correct navigation links', () => {
+  render(
+    <MemoryRouter>
+      <Header namePageTerm="Home Page" />
+    </MemoryRouter>
+  );
+
+  const homeLink = screen.getByRole('link', { name: /Home Page/i });
+  const aboutLink = screen.getByRole('link', { name: /About US/i });
+
+  expect(homeLink.getAttribute('href')).toBe('/');
+  expect(aboutLink.getAttribute('href')).toBe('/about');
 });
