+++ rs-week1/src/components/Header.css
@@ -1,4 +1,3 @@
-
 .header {
   background-color: #343a40;
   color: #fff;
@@ -21,7 +20,6 @@
 }
 
 .navigation-links>* {
-  /* margin-left: 0.5rem; */
   font-size: 18px;
   font-weight: bold;
   margin: 0.2rem;
@@ -47,6 +45,6 @@
   margin: 0.2rem;
 }
 
-.active{
+.active {
   color: #007bff;
 }
