+++ rs-week1/src/components/ProductCard.css
@@ -1,7 +1,7 @@
 .product-card {
   background-color: #ffffff;
   border: 1px solid #e2e2e2;
-  box-shadow: 0px 0px 6px rgba(0, 0, 0, 0.1);
+  box-shadow: 0px 0px 6px rgba(0, 0, 0, 0.5);
   display: flex;
   flex-direction: column;
   margin-bottom: 0.2rem;
@@ -45,7 +45,7 @@
   line-height: 1.5;
 }
 
-.product-card__info button {
+.product-card__button {
   padding: 10px;
   font-size: 16px;
   font-weight: bold;
@@ -54,6 +54,15 @@
   border: none;
   border-radius: 5px;
   cursor: pointer;
+  transition: all 0.1s ease-in-out;
+}
+
+.product-card__button:hover {
+  background-color: #0069d9;
+}
+
+.product-card__button:active {
+  background-color: rgb(101, 236, 83);
 }
 
 @media screen and (max-width: 1024px) {
