+++ rs-week1/src/components/pages/Home.css
@@ -5,6 +5,7 @@
   margin: 0 auto;
   max-width: 1500px;
   padding: 0 1rem;
+  gap: 8px;
 }
 
 .card-search {
