--- rs-week1/src/components/types/interface.js
@@ -1,1 +1,0 @@
-export {};
